fix: migrate middleware.ts to proxy.ts convention and fix TypeScript errors
