+++ packages/ui/package.json
@@ -57,6 +57,7 @@
     "vitest": "^4.0.18"
   },
   "peerDependencies": {
+    "next": "^15.0.0 || ^16.0.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "framer-motion": "^11.0.0 || ^12.0.0"
